requirements.txt
os
discord.py
